Add redirect functionality to social buttons
