+++ src/app/components/common/socialButtons/SocialButtons.js
@@ -3,6 +3,7 @@
 import { useContext } from 'react';
 import { BottomMessageBarContext } from '../../../context/bottomMessageBar/BottomMessageBarContext';
 import Image from 'next/image'
+import { redirectTo } from '@/app/utils/Utils';
 
 const socials = [
     {
@@ -25,24 +26,17 @@ const socials = [
 export default function SocialButtons() {
     const { showBottomMessageBar } = useContext(BottomMessageBarContext);
 
-    function redirectTo(url) {
-        showBottomMessageBar(
-            "Redirecting to " + url, 
-            "bg-green-700"
-        )
-
-        let timeout = setTimeout(() => {
-            window.open(url, '_blank');
-            clearTimeout(timeout);
-        }, 2000);
-    }
-
     return (
         <div className="flex flex-row justify-center items-center gap-6">
             {socials.map((social, index) => {
                 return (
                     <div key={index} className=" p-2 invert">
-                        <a onClick={() => redirectTo(social.href)}>
+                        <a onClick={
+                            () => {
+                                showBottomMessageBar(`Redirecting to ${social.name}`)
+                                redirectTo(social.href)
+                            }
+                            }>
                             <Image src={social.icon} alt={social.name} width={30} height={30} className='rounded-sm w-7 h-7 hover:opacity-70' />
                         </a>
                     </div>
