+++ src/app/utils/Utils.js
@@ -1,0 +1,7 @@
+export function redirectTo(url, wait=2000) {
+
+    let timeout = setTimeout(() => {
+        window.open(url, '_blank');
+        clearTimeout(timeout);
+    }, wait);
+}
